Small modification in the event when creating a new folder. Added a script to verify a user.
