+++ hardhat/hardhat.config.js
@@ -7,22 +7,10 @@ require('solidity-coverage');
 require('@nomiclabs/hardhat-etherscan');
 
 /**
- * Display the list of accounts.
- * @dev hh accounts
+ * Load tasks.
  */
-task('accounts', 'Prints the list of accounts', async () => {
-  const hdNode = ethers.utils.HDNode.fromMnemonic(process.env.WALLET_DEV_MNEMONIC);
-  const accounts = await ethers.getSigners();
-  let i = 0;
-  for (const account of accounts) {
-    const wallet = hdNode.derivePath(`m/44'/60'/0'/0/${i++}`);
-    const balance = await ethers.provider.getBalance(account.address);
-    console.log(
-      `- Address: ${account.address}: - Balance: ${ethers.utils.formatEther(balance)} ETH`,
-    );
-    console.log(`  Private Key: ${wallet.privateKey}`);
-  }
-});
+require('./tasks/accounts');
+require('./tasks/verifyUser');
 
 /** @type import('hardhat/config').HardhatUserConfig */
 module.exports = {
